warning fixes

+++ iphone/Classes/TiAppiOSUserDefaultsProxy.m
@@ -28,7 +28,7 @@
     return @"Ti.App.iOS.UserDefaults";
 }
 
--(void)_listenerAdded:(NSString*)type count:(int)count
+-(void)_listenerAdded:(NSString*)type count:(NSInteger)count
 {
 	if (count == 1 && [type isEqual:@"change"])
 	{
@@ -38,7 +38,7 @@
 	}
 }
 
--(void)_listenerRemoved:(NSString*)type count:(int)count
+-(void)_listenerRemoved:(NSString*)type count:(NSInteger)count
 {
 	if (count == 0 && [type isEqual:@"change"])
 	{
